relocating exp file, updating logic
